requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
groq
openai
python-multipart
fastapi
uvicorn
phi-agent
python-twitter
newsapi-python
ta-lib
pandas
matplotlib
logging
